requirements.txt
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
braceexpand==0.1.7
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.1
confection==0.1.4
contexttimer==0.3.3
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
decord==0.6.0
distlib==0.3.8
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.4
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.40
huggingface-hub==0.20.2
identify==2.5.33
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.29.2
ipython==8.19.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kaggle==1.6.3
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
numpy==1.26.3
omegaconf==2.3.0
opencv-python==4.9.0.80
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
plotly==5.18.0
portalocker==2.8.2
pre-commit==3.6.0
preshed==3.0.9
prettytable==3.9.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pycocoevalcap==1.2
pycocotools==2.0.7
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
safetensors==0.4.1
salesforce-lavis==1.0.2
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.11.4
sentencepiece==0.1.99
sentry-sdk==1.39.1
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
streamlit==1.30.0
tenacity==8.2.3
text-unidecode==1.3
thinc==8.2.2
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.0+cu116
torchaudio==0.12.0+cu116
torchvision==0.13.0+cu116
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.26.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
virtualenv==20.25.0
wandb==0.16.1
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.12
weasel==0.3.4
webdataset==0.2.86
webencodings==0.5.1
zipp==3.17.0
